Refactor: more consistent use of "worker" and "remote" names
A "remote" is an LXD remote that tests will run on. These remotes are
called "remote-ARCH". There is one per architecture.

A worker is a service (systemd service unit, or running instance of the
worker script) that pops jobs from the queues and runs them on the
remotes.

On top of the refactor there are some bug fixes and some cosmetic
changes.