worker: include architecture and release in the run_id hash
This will help preventing collisions, while still leaving the run_id
deterministic. To the best of my understanding, we don't need it to be
deterministic, it is just a nice property for it to have. This way the
hash is like a stamp representing what is being tested, and the same
test gets the same stamp.

In the future I see us using a fully random uuid instead of this, but
first we should stop using the run_id as a timestamp.